+++ src/components/SearchBar.vue
@@ -47,7 +47,7 @@
 
 <script>
 import { fuzzyMatcher } from "../utils/search";
-import enterprises from "../datas/enterprise.json";
+import enterprises from "../data/enterprise.json";
 import RightChevronIcon from "../assets/right-chevron.svg";
 import ResetIcon from "../assets/reset.svg";
