--- src/datas/enterprise.json
@@ -1,32 +1,0 @@
-[
-  {
-    "enterprise": "삼성전자",
-    "result": {
-      "aggressive": 10,
-      "confident": 9,
-      "responsible": 9,
-      "indivisual": 5,
-      "horizontal": 3
-    }
-  },
-  {
-    "enterprise": "카카오",
-    "result": {
-      "aggressive": 6,
-      "confident": 7,
-      "responsible": 7,
-      "indivisual": 8,
-      "horizontal": 9
-    }
-  },
-  {
-    "enterprise" : "LG CNS",
-    "result" : {
-      "aggressive" : 7,
-      "confident" : 7,
-      "responsible": 7,
-      "indivisual": 7,
-      "horizontal": 7
-    }
-  }
-]
